+++ backend/src/main/java/com/techsage/banking/services/interfaces/BankAccountService.java
@@ -7,8 +7,8 @@ import java.util.List;
 
 public interface BankAccountService {
     List<BankAccountDto> getAll();
-    BankAccount getById(int id);
+    BankAccount getById(long id);
     BankAccount create(BankAccount bankAccount);
     BankAccount update(BankAccount bankAccount);
-    void delete(int id);
+    void delete(long id);
 }
