+++ backend/src/main/java/com/techsage/banking/services/interfaces/TransactionService.java
@@ -2,13 +2,14 @@ package com.techsage.banking.services.interfaces;
 
 import com.techsage.banking.models.Transaction;
 import com.techsage.banking.models.dto.TransactionDto;
+import org.springframework.stereotype.Service;
 
 import java.util.List;
 
 public interface TransactionService {
     List<TransactionDto> getAll();
-    Transaction getById(int id);
+    Transaction getById(long id);
     Transaction create(Transaction transaction);
     Transaction update(Transaction transaction);
-    void delete(int id);
+    void delete(long id);
 }
